Add files via upload
this is a second version of the guessing game I have already created.
the concept of the game: the player has 3 chances to guess the number the computer has chosen (providing that the number is between 1 and 10). the computer will provide hints to the player for every wrong answer to help them make a good guess next time.